+++ srcs/editor_loop/editor_linedef.c
@@ -105,13 +105,13 @@ void	fill_abscissa_ordinate(t_map_editor *map, t_dot mouse, int gap)
 		map->ordinate = (t_line){(t_dot){0, 0}, (t_dot){0, 0}};
 }
 
-SDL_bool		is_line_horizontally(t_dot lp1, t_dot lp2, double angle)
+SDL_bool		is_line_horizontally(t_dot lp1, t_dot lp2, float angle)
 {
 	t_dot	p1;
 	t_dot	p2;
 	int		dx;
 	int		dy;
-	double	line_angle;
+	float	line_angle;
 
 	if (lp1.x == lp2.x)
 		return (SDL_FALSE);
@@ -133,12 +133,12 @@ SDL_bool		is_line_horizontally(t_dot lp1, t_dot lp2, double angle)
 	}
 	dy = p2.y - p1.y;
 	dx = p2.x - p1.x;
-	line_angle = atan(dy/(double)dx) * 180 / M_PI;
+	line_angle = atan(dy/(float)dx) * 180 / M_PI;
 	return ((line_angle < angle) ? SDL_TRUE : SDL_FALSE);
 }
 
 
-SDL_bool		is_line_vertically(t_dot lp1, t_dot lp2, double angle)
+SDL_bool		is_line_vertically(t_dot lp1, t_dot lp2, float angle)
 {
 	t_dot	p1;
 	t_dot	p2;
